requirements.txt
botocore==1.34.71
boto3==1.34.71
PyMuPDF==1.24.0
streamlit==1.32.2
pandas==2.2.1
PyYAML==6.0.1
